添加参数验证异常处理

+++ data/src/main/java/model/Shelf.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 public class Shelf implements Serializable{
     private String id;
     private Long bookId;
-    private BigInteger userId;
+    private Long userId;
     private LocalDateTime addDate;
     private LocalDateTime updateDate;
     private Integer state;
